feat(vector): add scaleBy to scale
