+++ src/algebra/vector/scale.spec.ts
@@ -1,4 +1,4 @@
-import { scale } from './scale';
+import { scale, scaleBy } from './scale';
 
 type Vector = number[];
 
@@ -16,5 +16,28 @@ describe('algebra/vector/scale', () => {
     const result = scale(v1, scalar);
 
     expect(result).toEqual(components.map(c => c * scalar));
+  });
+
+  describe('scale-by', () => {
+    it('should scale by 0', () => {
+      const scaleByZero = scaleBy(0);
+
+      const result = scaleByZero([1, 2, 3]);
+      expect(result).toEqual([0, 0, 0]);
+    });
+
+    it('should scale by 1', () => {
+      const scaleByZero = scaleBy(1);
+
+      const result = scaleByZero([1, 2, 3]);
+      expect(result).toEqual([1, 2, 3]);
+    });
+
+    it('should scale by 2', () => {
+      const scaleByZero = scaleBy(2);
+
+      const result = scaleByZero([1, 2, 3]);
+      expect(result).toEqual([2, 4, 6]);
+    });
   });
 });
